Explicitly use bash to fix "bad substitution" error
When I run the `nsenter-node.sh` script without this change, I get the following error:
```
/home/agr/.bin/nsenter-node.sh: 9: Bad substitution
```
I think that's because `/bin/sh` is not bash on my system. So let's explicitly declare in the shebang that we use bash. Also use the recommended `/usr/bin/env` to find the path to bash.